+++ src/matcher_inner/debug.rs
@@ -1,4 +1,4 @@
-use crate::matrix::{MatrixCell, ScoreCell};
+use crate::matcher_inner::matrix::{MatrixCell, ScoreCell};
 use std::fmt::{Debug, Formatter, Result};
 
 impl Debug for ScoreCell {